xrandr: check_strtod should return double, not int
check_strtod performs error checking around strtod to ensure that
arguments are correctly processed. However, it also accidentally(?)
cast the result to int, which was then universally cast back to double
by all callers. Narrowing and re-widening the type doesn't make any sense.

+++ xrandr.c
@@ -2010,7 +2010,7 @@ check_strtol(char *s)
     return result;
 }
 
-static int
+static double
 check_strtod(char *s)
 {
     char *endptr;
